Add .phpstorm.meta.php
As future replacement of dynamicReturnTypeMeta.json because the meta is
"PhpStorm Native" while the dynamic return types plugin is no longer actively
maintained and half broken.

Bonus is, as this comes into your vendor dir, PhpStorm picks up on it, so
projects using Parable Di don't have to specify it themselves.